requirements.txt
openai>=1.0.0
faiss-cpu>=1.7.0
numpy>=1.24.0
python-dotenv>=1.0.0
boto3>=1.34.0
pinecone-client>=3.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
httpx>=0.24.0
datasets>=2.14.0
